// src/pages/hr/employees/create/data/interface.ts
export interface IEmployee {
  address?: string;
  city?: string;
  cnic?: string;
  companyEmail?: string;
  contactDetail?: string;
  createdAt?: string;
  createdBy?: string;
  dateOfBirth?: string;
  departmentId?: string;
  designationId?: string;
  email?: string;
  employeeDepartment?: string;
  employeeDesignation?: string;
  employeeId?: string;
  fullName?: string;
  gender?: string;
  id?: number;
  isDeleted?: boolean;
  joiningDate?: string;
  managerId?: string;
  state?: string;
  status?: string;
  updatedAt?: string;
  updatedBy?: string;
  workType?: string;
}